+++ NKDBarcode/Classes/NKDUPCABarcode.m
@@ -79,7 +79,7 @@
 	char *    code = (char *) [content cStringUsingEncoding:NSStringEncodingConversionAllowLossy];
 	if (strlen(code) == 11)
 	{
-		i = strlen(code);
+		i = (int)strlen(code);
 		while (i-- > 0) {
 			if (even) evenSum += code[i]-'0';
 			else      oddSum += code[i]-'0';
